+++ example/.eslintrc.cjs
@@ -31,4 +31,5 @@ module.exports = {
     ],
     "@typescript-eslint/no-explicit-any": "warn",
   },
+  ignorePatterns: ["dist"],
 };