+++ example/src/data/mutations/LoginMutation/useLoginMutation.mocks.gen.ts
@@ -1,0 +1,5 @@
+import { useLoginMutation } from "./useLoginMutation";
+
+type QueryData = ReturnType<typeof useLoginMutation>["data"];
+
+export const mockLoginMutationData: QueryData = undefined; // TODO: Replace with mock mutation data of type: LoginData