+++ example/src/data/queries/useReportsQuery.mocks.gen.ts
@@ -1,0 +1,5 @@
+import { useReportsQuery } from "./useReportsQuery";
+
+type QueryData = ReturnType<typeof useReportsQuery>["data"];
+
+export const mockReportsQueryData: QueryData = undefined; // TODO: Replace with mock query data of type: ReportsQueryData