+++ example/src/otherQueries/useAssetsQuery.mocks.gen.ts
@@ -1,0 +1,5 @@
+import { useAssetsQuery } from "./useAssetsQuery";
+
+type QueryData = ReturnType<typeof useAssetsQuery>["data"];
+
+export const mockAssetsQueryData: QueryData = undefined; // TODO: Replace with mock query data of type: AssetsQueryData