+++ example/src/otherQueries/useLogoutMutation.mocks.gen.ts
@@ -1,0 +1,5 @@
+import { useLogoutMutation } from "./useLogoutMutation";
+
+type QueryData = ReturnType<typeof useLogoutMutation>["data"];
+
+export const mockLogoutMutationData: QueryData = undefined; // TODO: Replace with mock mutation data of type: D