+++ src/data-generation/index.ts
@@ -1,0 +1,6 @@
+import { generateMockDataFiles } from "./generator";
+import { parseQueryMutationFiles } from "./parser";
+import { MockDataConfig } from "./types";
+
+export { generateMockDataFiles, parseQueryMutationFiles };
+export type { MockDataConfig };